+++ resources/js/Components/Items/ApplicationLogo.vue
