+++ resources/js/Layouts/GuestLayout.vue
@@ -1,5 +1,5 @@
 <script setup>
-import ApplicationLogo from '@/Components/ApplicationLogo.vue';
+import ApplicationLogo from '@/Components/Items/ApplicationLogo.vue';
 import { Link } from '@inertiajs/vue3';
 </script>
