+++ resources/js/Pages/Profile/Partials/DeleteUserForm.vue
@@ -1,10 +1,10 @@
 <script setup>
-import DangerButton from '@/Components/DangerButton.vue';
-import InputError from '@/Components/InputError.vue';
-import InputLabel from '@/Components/InputLabel.vue';
-import Modal from '@/Components/Modal.vue';
-import SecondaryButton from '@/Components/SecondaryButton.vue';
-import TextInput from '@/Components/TextInput.vue';
+import DangerButton from '@/Components/Items/DangerButton.vue';
+import InputError from '@/Components/Items/InputError.vue';
+import InputLabel from '@/Components/Items/InputLabel.vue';
+import Modal from '@/Components/Items/Modal.vue';
+import SecondaryButton from '@/Components/Items/SecondaryButton.vue';
+import TextInput from '@/Components/Items/TextInput.vue';
 import { useForm } from '@inertiajs/vue3';
 import { nextTick, ref } from 'vue';
