+++ resources/js/Pages/Profile/Partials/UpdateProfileInformationForm.vue
@@ -1,8 +1,8 @@
 <script setup>
-import InputError from '@/Components/InputError.vue';
-import InputLabel from '@/Components/InputLabel.vue';
-import PrimaryButton from '@/Components/PrimaryButton.vue';
-import TextInput from '@/Components/TextInput.vue';
+import InputError from '@/Components/Items/InputError.vue';
+import InputLabel from '@/Components/Items/InputLabel.vue';
+import PrimaryButton from '@/Components/Items/PrimaryButton.vue';
+import TextInput from '@/Components/Items/TextInput.vue';
 import { Link, useForm, usePage } from '@inertiajs/vue3';
 
 defineProps({
